+++ src/app/modules/trade/component/trade-settings/trade-settings.component.ts
@@ -1,7 +1,8 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { AudioFile, AudioService } from '@app/audio';
+import { EnumValues } from '@app/enum';
 import { FeatureSettingsComponent } from '@app/feature';
-import { TradeFeatureSettings } from '@modules/trade/trade-feature-settings';
+import { TradeFeatureSettings, TradeFilter, TradeLayout } from '@modules/trade/trade-feature-settings';
 
 @Component({
   selector: 'app-trade-settings',
@@ -11,6 +12,9 @@ import { TradeFeatureSettings } from '@modules/trade/trade-feature-settings';
 })
 export class TradeSettingsComponent extends FeatureSettingsComponent<TradeFeatureSettings> {
   public displayWithVolume = (volume: number) => `${volume}%`;
+
+  public layouts = new EnumValues(TradeLayout);
+  public filters = new EnumValues(TradeFilter);
 
   constructor(private readonly audio: AudioService) {
     super();